simulation2/GabibbiTowers/backend/requirements.txt
fastapi==0.97.0
fastapi-login==1.9.0
pydantic==1.10.9
uvicorn==0.22.0
pycryptodome==3.16.0
requests==2.22.0